+++ generator/src/lib.rs
@@ -1905,7 +1905,7 @@ fn derive_default(
     Some(q)
 }
 
-fn derive_send_sync(struct_: &vkxml::Struct) -> Option<TokenStream> {
+fn derive_send_sync(struct_: &vkxml::Struct, has_lifetime: bool) -> Option<TokenStream> {
     if !struct_
         .elements
         .iter()
@@ -1916,9 +1916,10 @@ fn derive_send_sync(struct_: &vkxml::Struct) -> Option<TokenStream> {
         return None;
     }
     let name = name_to_tokens(&struct_.name);
+    let lifetime = has_lifetime.then(|| quote!(<'_>));
     let q = quote! {
-        unsafe impl Send for #name <'_> {}
-        unsafe impl Sync for #name <'_> {}
+        unsafe impl Send for #name #lifetime {}
+        unsafe impl Sync for #name #lifetime {}
     };
     Some(q)
 }
@@ -2590,7 +2591,7 @@ pub fn generate_struct(
 
     let debug_tokens = derive_debug(struct_, &members, union_types, has_lifetime);
     let default_tokens = derive_default(struct_, &members, has_lifetime);
-    let send_sync_tokens = derive_send_sync(struct_);
+    let send_sync_tokens = derive_send_sync(struct_, has_lifetime);
     let setter_tokens = derive_getters_and_setters(struct_, &members, root_structs, has_lifetimes);
     let manual_derive_tokens = manual_derives(struct_);
     let dbg_str = if debug_tokens.is_none() {
@@ -3249,6 +3250,14 @@ pub fn write_source_code<P: AsRef<Path>>(vk_headers_dir: &Path, src_dir: P) {
             s.elements
                 .iter()
                 .filter_map(get_variant!(vkxml::StructElement::Member))
+                .filter(|x| {
+                    x.name.as_deref() == Some("pNext")
+                        // Opaque types (typically external system handles) are treated as pointers
+                        // rather than borrows. They don't carry a lifetime, unless they are wrapped
+                        // in a slice which inherently is lifetime-bounded.
+                        || x.size.is_some()
+                        || !is_opaque_type(&x.basetype)
+                })
                 .any(|x| x.reference.is_some())
         })
         .map(|s| name_to_tokens(&s.name))